+++ feature_browse/src/main/java/ru/kggm/feature_browse/presentation/ui/characters/details/CharacterDetailsViewModel.kt
@@ -25,7 +25,7 @@ class CharacterDetailsViewModel @Inject constructor(
     }
 
     fun loadCharacter(id: Long) = safeLaunch {
-        getCharacterById(id).let {
+        getCharacterById(id.toInt()).let {
             characterFlow.tryEmit(it.toPresentationEntity())
         }
     }
